fixed bad egg in daycare
giving pokemon to the daycare was writing to var_0x8005. the interview easychat screen also uses that var. This was causing weirdness that led to the player receiving an odd egg after giving a machop and eevee to the daycare and then entering abra 4x in the easychat.

Now, the var gets wiped before calling the easychat.