Fixed bug contest party removal bug (i hope)
This only impacted players who used speed up, which behaves differently on different emulators.

Possible cause 1: Eastern gate guard does not appear faster than the player could walk right.
Solution 1: Move the player further left, so it will take twice as long to hit the warp. Not possible now in my testing.

PC2: Player was unlocked due to the release at the end of the gate script
Solution 2: removed the release after the warp, and moved the setvar to before the warp so there is no possibility of a hanging release.

Added dialogue about exp removal.